Modified the trusty download to use testing instead the sources.list
The issue is that the trusty download is available for rabbitmq-server,
but not for erlang-nox. If trusty is used, installing rabbitmq-server
is impossible as it requires the new erlang dependencies.

Solution is to check if the version is trusty, if so just use the
"testing" list that travis and the team has used in the past.

Another change is adding another line for the deb list, erlang is in
https://dl.bintray.com/rabbitmq-erlang/debian, not
https://dl.bintray.com/rabbitmq/debian.